Add property and related to it entities controllers

+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/models/PropertyAvailability.java
@@ -17,7 +17,7 @@ public class PropertyAvailability {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Integer id;
 
     private LocalDate bookedSince;
